refactor(arch): VfsStrategy を実バックエンド語彙に整合し sqliteEngineContext 別名を削除
未実装の案B（opfs-async-main）を VfsStrategy から削除し、OPFS ディレクトリ
有り・sync 無し環境が resolver と同じく idb を指すように整合（診断表示の
実バックエンド一致）。薄い別名 sqliteEngineContext.ts（15 行）を削除し、
18 ファイルの import（vi.mock・動的 import 含む）を sqliteEngineHost.js 直参照に
付け替え、SqliteEngineContext 型名も SqliteEngineHost に統一。振る舞い変更なし。
スパイク（dig-findings-2026-09-05）の PBI-C1。
