+++ src/offscreen/OpfsWorkerBackend.ts
@@ -1,11 +1,10 @@
 // src/offscreen/OpfsWorkerBackend.ts
 import type { SqliteEngineHost } from './sqliteEngineHost.js';
-import type { SqliteEngineContext } from './sqliteEngineContext.js';
 import type { StorageBackend, InsertResult, InsertBatchResult, QuerySearchResult, MutationResult, StarResult, PurgeResult, FtsSizeResult, BackupResult, CountResult, HealthResult, AuditLogQueryResult, StatusResult, BackendOrError } from './StorageBackend.js';
 import type { BrowsingLogRecord, BrowsingLogEntry, StorageQuery, AuditLogRecord, AuditLogEntry } from '../utils/sqlite-types.js';
 
 export class OpfsWorkerBackend implements StorageBackend {
-  constructor(private engine: SqliteEngineHost | SqliteEngineContext) {}
+  constructor(private engine: SqliteEngineHost) {}
 
   async insert(record: BrowsingLogRecord): Promise<BackendOrError<InsertResult>> {
     const result = await this.engine.sendToOpfsWorker('INSERT', record) as { id: number };
