+++ src/offscreen/backendResolver.ts
@@ -14,7 +14,6 @@
 import type { OpfsCapabilities } from './opfsCapabilities.js';
 import { detectLiveVfsStrategy } from './opfsCapabilities.js';
 import type { SqliteEngineHost } from './sqliteEngineHost.js';
-import type { SqliteEngineContext } from './sqliteEngineContext.js';
 import type { StorageBackend } from './StorageBackend.js';
 import { NoopBackend } from './StorageBackend.js';
 
@@ -45,7 +44,7 @@ export function resolveBackend(state: PostInitState): BackendType {
  * Falls back to NoopBackend if the resolved type has no matching adapter.
  */
 export async function createBackend(
-  context: SqliteEngineHost | SqliteEngineContext,
+  context: SqliteEngineHost,
   resolved: BackendType
 ): Promise<StorageBackend> {
   switch (resolved) {
